+++ app/Http/Controllers/ClassesController.php
@@ -5,14 +5,15 @@ namespace App\Http\Controllers;
 use App\Models\Classes;
 use App\Models\ClassRoom;
 use App\Models\Teacher;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Contracts\Support\Renderable;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
 class ClassesController extends Controller {
     const ROUTE = 'http://127.0.0.1:8000/';
+    const ADMIN = 'admin';
 
     /**
      * Display a listing of the resource.
@@ -82,8 +83,11 @@ class ClassesController extends Controller {
      * @return RedirectResponse
      */
     public function destroy($code):RedirectResponse {
-        (new Classes())->where('code', $code)->delete();
-        return redirect()->away(self::ROUTE.'class/show')->with('error', 'Clase eliminada correctamente')->with('classes', Classes::all());
+        if(Auth::check() && Auth::user()->name === self::ADMIN) {
+            (new Classes())->where('code', $code)->delete();
+            return redirect()->away(self::ROUTE.'class/show')->with('error', 'Clase eliminada correctamente')->with('classes', Classes::all());
+        }
+        return $this->redirectToHome('error', 'Solo el Administrador puede realizar esta operacion');
     }
 
     private function redirectToHome($type, $message): RedirectResponse {
